Updated to work with Google Apps Directory Sync.
Modified the installer to incorporate the settings needed for GADS to
function correctly, however the update feature is not working, so you
must do a uninstall then reinstall currently.  password_update.sh now
modifies the users LDAP record with a SHA hash of the password so GADS
can push this to Google Apps, make sure you use DACLs when using this
setting.   crypto.sh now also handles the SHA hashing.  Checked with
10.5 / 10.6 / 10.7 and confirmed working.